Use autogen v0.4 explicitly for stability
Also moved the checkout line closer to where it's used.

This script was using the latest version of autogen, which
I am planning to rewrite in the future to be written in Go,
at which point what is you're doing here will not work.

Keeping the checkout to be fixed to a specific version will
enable you to not be broken, and to switch this to the new
form when you're ready.